프로젝트 정리 낮은 수준의 경고 무시 pragma 추가 자잘한 경고발생 코드, 안나도록 수정.

+++ source/OliveLib/include/OvMesh.h
@@ -26,9 +26,9 @@ public:
 		SRenderData():faceStream(NULL), vertDecl(NULL), vertexCount(0), faceCount(0){};
 		SVertexStreamInfo	vertStreamInfo[ STAGE_MAX ];
 		LPDIRECT3DVERTEXDECLARATION9 vertDecl;
-		size_t	vertexCount;
+		UINT	vertexCount;
 		LPDIRECT3DINDEXBUFFER9	faceStream;
-		size_t	faceCount;
+		UINT	faceCount;
 	};
 
 	OvMesh();
