+++ source/OliveLib/include/OvMeshLoaderBase.h
@@ -11,7 +11,7 @@ protected:
 		, SVertexStreamInfo& streamMedium
 		, SVertexStreamInfo& streamHigh
 		, LPDIRECT3DVERTEXDECLARATION9 declVert
-		, size_t countVert
+		, UINT countVert
 		, LPDIRECT3DINDEXBUFFER9 streamFace
-		, size_t countFace );
+		, UINT countFace );
 };
