+++ source/OliveLib/include/OvRenderingCommon.h
@@ -19,12 +19,12 @@ struct SVertexStreamInfo
 	SVertexStreamInfo(): vertexStream( NULL ), vertexStride( 0 ), streamOffset( 0 ){};
 	SVertexStreamInfo( 
 		LPDIRECT3DVERTEXBUFFER9 _vertexStream,
-		size_t	_vertexStride,
-		size_t	_streamOffset)
+		UINT	_vertexStride,
+		UINT	_streamOffset)
 		: vertexStream( _vertexStream )
 		, vertexStride( _vertexStride )
 		, streamOffset( _streamOffset ){};
 	LPDIRECT3DVERTEXBUFFER9 vertexStream;
-	size_t	vertexStride;
-	size_t	streamOffset;
+	UINT	vertexStride;
+	UINT	streamOffset;
 };
